build working for node, but I still need to tweak it for the web tool
